Actualizacion Archivo Algoritmos Luis Taque
