+++ .vscode/launch.json
@@ -1,0 +1,13 @@
+{
+    // Use IntelliSense para saber los atributos posibles.
+    // Mantenga el puntero para ver las descripciones de los existentes atributos.
+    // Para más información, visite: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "APPS-MOBILE-DART",
+            "request": "launch",
+            "type": "dart"
+        }
+    ]
+}
