--- test/poo_test.dart
@@ -1,8 +1,0 @@
-import 'package:poo/poo.dart';
-import 'package:test/test.dart';
-
-void main() {
-  test('calculate', () {
-    expect(calculate(), 42);
-  });
-}
